+++ src/commands/index.ts
@@ -4,6 +4,7 @@ import Help from './help';
 import BaseCommand from './base';
 import About from './about';
 import Uptime from './uptime';
+import Roll from './roll';
 
 export interface CommandMap {
     [key: string]: BaseCommand;
@@ -15,4 +16,5 @@ export default {
     help: new Help(),
     about: new About(),
     uptime: new Uptime(),
+    roll: new Roll(),
 } as CommandMap;