+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Changelog
 
-## [Unreleased]
+## [v0.2.0]
 
+* Fixed: Fixed where it was possible to create a simple command using a built-in command's name (e.g. `!command add command ...`)
 * Fixed: Exception when a command was sent to the bot via DM.
 * Added: Reminder system via the `!reminders` command.
 * Added: If a subcommand is missing, help will be displayed for that particular command.
 * Added: The `!help` command now shows commands with available help.
-* Added: `!about` and `!uptime` commands.
+* Added: The help output for commands includes the command prefix.
+* Added: `!about`, `!uptime` and `!roll` commands.
 
 ## [v0.1.0]